Major Changes: BedExtruderWigdet
- Implement ui
- Implement methods, signals and connections in source

I don't think we will need to implement other way to
show the extruder temperature, since the GCode Wiki
says that the temperature returned is of the current
extruder. So this widget will fill it's purpose.

Info about M105:
"Request the temperature of the current extruder and the build
base in degrees Celsius"
http://reprap.org/wiki/G-code#M105:_Get_Extruder_Temperature

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>