Remove connection related to the GCodeDock
This connection was making atelier crash, so I moved to the
openFile method the include of the KTextEditor extras.
It makes sense to only show the extras when a file is open.
I have an idea on removal of the extras, but it will need
more work on the GCodeEditor itself.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>