Update CMakeLists
Updates necessary to set Qt and KDE Frameworks base
version to be used on this project.

Add a config.h.cmake file to make the version be
set automatically based on this file. This
file gets the version from the CMakeLists
version. And I think that those variables can be
manage by scripts, based on Kate setup.

Signed-off-by: Lays Rodrigues <laysrodriguessilva@gmail.com>