Add an axis gnomon entity to the scene
It reuses the per vertex color fragment shader but replaces the vertex
shader with our own to make sure we're tracking the camera orientation
and position the object relatively to the camera.

Note this doesn't play well with frustum culling (since the computation
is wrong in that case) hence why it's turned off. Could be reenabled
later if needed by providing our own two passes framegraph and having
the axis gnomon rendered in a separate pass.

Differential Revision: https://phabricator.kde.org/D15045